Add files via upload
This is a script I wrote that processes microarray data.The script  first scales and organizes the data(expression levels and probes). Then differential expression is performed via Limma. With the use of a random forest model the differentially expressed genes are sorted based variable importance  The  probes of the  20 most important genes are then mapped to gene symbols. That information is then output as a CSV file for further analysis. Next , another random forest model is prepared using those genes and confusion matrix and roc curve are output. The  data is the Golub et al dataset. However, another expression set could be processed in this way.